smoothing: rfctr counts smoothing, move test into separate file, smoothing for col percent
